Change in a lot of files
Kasper and Anders changed the interaction with the cook so the window, button and text changed depending on the situation.  Fixed a bug so you now need to be in the same room as the cook to talk with him. FInally we also checked every file and wrote missing javadocs, removed spaces and unnecessary lines.